Move even more code around
