+++ packages/frontend/src/components/chart/configure-old/getOutputs.ts
