+++ packages/frontend/src/components/chart/configure-old/index.ts
@@ -1,0 +1,31 @@
+import { getOutputs } from './getOutputs'
+import { initHover } from './hover'
+import { plot } from './plot'
+import { ChartStateWithInput, makeChartState } from './state'
+import { toUiState, UiState } from './ui'
+
+export function configureChart() {
+  const chart = document.querySelector<HTMLElement>('[data-role="chart"]')
+  if (!chart) {
+    return
+  }
+
+  const outputs = getOutputs(chart)
+
+  let uiState: UiState | undefined
+  const chartState = makeChartState(chart, () => update())
+  window.addEventListener('resize', () => requestAnimationFrame(draw))
+
+  function update() {
+    uiState =
+      chartState.mainInput && toUiState(chartState as ChartStateWithInput)
+    requestAnimationFrame(draw)
+  }
+
+  const hover = initHover(chart)
+
+  function draw() {
+    hover.update(uiState)
+    plot(uiState, outputs)
+  }
+}
