+++ packages/frontend/src/components/chart/configure-old/state/getControls.ts
