+++ packages/frontend/src/components/chart/configure-old/state/getEndpoint.ts
