+++ packages/frontend/src/components/chart/configure-old/state/getType.ts
