+++ packages/frontend/src/components/chart/configure/controls/onCheckboxChange.ts
