+++ packages/frontend/src/components/chart/configure/effects/effects.ts
