+++ packages/frontend/src/components/chart/configure/render/fillBelowChart.ts
