+++ packages/frontend/src/components/chart/configure/render/strokeChartLine.ts
