+++ packages/frontend/src/components/chart/configure/render/style.ts
