+++ packages/frontend/src/components/chart/configure/state/empty.ts
