+++ packages/frontend/src/components/chart/configure/update/updateControlsChanged.ts
@@ -8,7 +8,7 @@ import {
   TokenChangedMessage,
 } from '../messages'
 import { State } from '../state/State'
-import { calculateView } from './calculateView'
+import { calculateView } from './view/calculateView'
 
 export function updateControlsChanged(
   state: State,
