+++ packages/frontend/src/components/chart/configure/update/updateShowAlternativeTvlChanged.ts
@@ -1,7 +1,7 @@
 import { Effect } from '../effects/effects'
 import { ShowAlternativeTvlChangedMessage } from '../messages'
 import { State } from '../state/State'
-import { calculateView } from './calculateView'
+import { calculateView } from './view/calculateView'
 
 export function updateShowAlternativeTvlChanged(
   state: State,
