+++ packages/frontend/src/components/chart/configure/update/view/calculateTicks.ts
