+++ packages/frontend/src/components/chart/configure/update/view/format.ts
@@ -1,4 +1,4 @@
-import { formatLargeNumber } from '../../../../utils'
+import { formatLargeNumber } from '../../../../../utils'
 
 export function formatCurrency(value: number, currency: string) {
   const num = formatLargeNumber(value)
