+++ packages/frontend/src/components/chart/configure/update/view/getHoverIndex.ts
@@ -1,0 +1,6 @@
+export function getHoverIndex(mouseX: number | undefined, points: number) {
+  if (mouseX && points > 0) {
+    // This only works if the points are evenly distributed
+    return Math.round(mouseX * (points - 1))
+  }
+}
